dev: Add checks for existing symlinks in 'run prepare'
The current 'run prepare' script uses sudo to create the
symlinks for python and nodejs. For many systems, this is
unnecessary and causes the user to be presented a request
for password. Additionally, this could affect any previous
system configuration. This change makes the symlinks only
be created if they don't already exist.