dev: autotester false positive when exception in single module test
The autotester browser UI had a bug where if the test consisted of one
module and it threw an exception, then the test would report a false
positive. This commit fixes that problem by changing the way that
'compare' operates. Additionally, now when an exception is thrown
the test that threw this exception is immediately obvious.