modules: Initial implementation of the re module.
This commit contains the initial implementation of the python regular
expressions module. This includes a module with two basic modes. First
mode attempts to replicate the python re module's regular expression
syntax by translating it to a javascript compatible syntax. The second
is a strictly javascript only regular expression syntax.

This commit also includes an initial set of unit tests for the re module
running in both modes. Not all methods are implemented yet. This commit
is intended to provide a model for review.