modules: Initial implementation of the `logging` python module
This commit includes an initial implementation of the logging module. It
includes a basic unit test for the logger only. The `logging.basicConfig`
and `logging.config.dictConfig` methods are not tested at this point.

I've stubbed in placeholders for some additional logging.Handler objects
that I think could be useful in the future but these are not tested
yet.