libs: Initial implementation of python's `warnings` module
This commit contains an initial implementation of the warnings module
for python. Note that the `catch_warnings` method is not implemented.
For now it is suggested that the user use the `warn_explicit` method
as the `warn` method does not do a good job of providing contextual
information.

This commit also updates the logging module to add back in the ability
to push warning messages through the logging interface.