core: Fixed bug in the repr method for dict and list-like objects
The List and Dict object types were using repeated code from
repr to implement the stringification of the objects. This was causing both types to misrepresent objects that contained 'None' or 'null' values. This
commit changes to using a recursive invokation of the repr function to
reduce code duplication and cause the unit tests for repr/str to pass