core: Fixed bugs in the way that dict objects are created.
This commit primarily targets issue #221 which discusses a bug in the
the shallow copy operation of dict() on existing dict. While addressing
this bug I also noticed some other issues, specifically with how the
dict() methods error handling does not raise exceptions for invalid args.
Additionally, the commit addresses the case where a user passes a
hierarchical js object to be converted by the dict() method by using a
recursive call on the dict() method.