compiler: Added __name__ macro feature similar to __file__
This commit adds a new macro that implements the `__name__` feature of
the python language. When the compiler encounters `__name__` in the
python source file, the __name__ is replaced with the name of the
python module as a string. This is a compile-time substitution similar
to the `__file__` and `__line__` parameters.