module:logging: Debugged core feature set of the logging module.
This commit includes a core set of features for the logging module
and a set of basic unit tests covering the major use cases. Note that
there are still some rough edges here and there. Currently, the only
available handler is the StreamHandler to the console.log, but the
Buffering/Memory handlers also have tested implementations.