example support for lexical operator overloading mode
A possible method for avoiding the overhead of `__call__` on overloaded operators.  Adds an optional argument that does a compile-time swap for the pythonic overload methods instead of the full call search.  So where this python:

    __pragma___('opov')
   a += b
   __ pragma__ ('noopov')

generates this JS:

     	var a= __call__ (__iadd__, null, a, b);

with the change this version:

    __pragma___('opov', 'fast')
   a += b
   __ pragma__ ('noopov')

generates
    
    a.__iadd__(b)

The intent would be to use this fairly narrowly in performance sensitive code.  

As written the following would fail if the left-hand operand does not have the correct magic methods.  For user code (especially pythong) that seems fine; it's an annoyance for native types.   However adding the magic methods to `Number.prototype` works, although I did not include it here because I was not sure where it would best be shimmed in